pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lordcapulet"
version = "0.1.0"
description = "AiiDA plugins for constrained DFT+U calculations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Alberto Carta", email = "your.email@example.com"},
]
keywords = ["aiida", "dft", "quantum-espresso", "constrained", "magnetism"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "aiida-core>=2.0.0",
    "aiida-quantumespresso>=4.0.0",
    "numpy",
]

[project.urls]
Homepage = "https://github.com/alberto-carta/aiida-LordCapulet"
Repository = "https://github.com/alberto-carta/aiida-LordCapulet"
Issues = "https://github.com/alberto-carta/aiida-LordCapulet/issues"

[project.entry-points."aiida.calculations"]
"lordcapulet.constrained_pw" = "lordcapulet.calculations.constrained_pw:ConstrainedPWCalculation"

[project.entry-points."aiida.tools.calculations"]
"lordcapulet.constrained_pw" = "aiida_quantumespresso.tools.calculations.pw:PwCalculationTools"

[project.entry-points."aiida.workflows"]
"lordcapulet.afm_scan" = "lordcapulet.workflows.afm_scan:AFMScanWorkChain"
"lordcapulet.constrained_scan" = "lordcapulet.workflows.constrained_scan:ConstrainedScanWorkChain"

[tool.setuptools.packages.find]
include = ["lordcapulet*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
